refactor(api): move findPaginatedFilteredUsers controller method
in src/identity-access-management
