+++ api/src/identity-access-management/application/user/user.admin.controller.js
@@ -1,5 +1,22 @@
+import { usecases as libUsecases } from '../../../../lib/domain/usecases/index.js';
 import { usecases } from '../../domain/usecases/index.js';
+import * as userForAdminSerializer from '../../infrastructure/serializers/jsonapi/user-for-admin.serializer.js';
 import * as userLoginSerializer from '../../infrastructure/serializers/jsonapi/user-login-serializer.js';
+
+/**
+ *
+ * @param request
+ * @param h
+ * @param {object} dependencies
+ * @param {UserForAdminSerializer} dependencies.userForAdminSerializer
+ * @returns {Promise<*>}
+ */
+const findPaginatedFilteredUsers = async function (request, h, dependencies = { userForAdminSerializer }) {
+  const { filter, page } = request.query;
+
+  const { models: users, pagination } = await libUsecases.findPaginatedFilteredUsers({ filter, page });
+  return dependencies.userForAdminSerializer.serialize(users, pagination);
+};
 
 /**
  * @param request
@@ -16,8 +33,9 @@ const unblockUserAccount = async function (request, h, dependencies = { userLogi
 
 /**
  * @typedef {object} UserAdminController
+ * @property {function} findPaginatedFilteredUsers
  * @property {function} unblockUserAccount
  */
-const userAdminController = { unblockUserAccount };
+const userAdminController = { findPaginatedFilteredUsers, unblockUserAccount };
 
 export { userAdminController };
