+++ api/tests/unit/application/users/user-controller_test.js
@@ -4,7 +4,6 @@ import { DomainTransaction } from '../../../../lib/infrastructure/DomainTransact
 import { usecases as devcompUsecases } from '../../../../src/devcomp/domain/usecases/index.js';
 import { evaluationUsecases } from '../../../../src/evaluation/domain/usecases/index.js';
 import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../src/identity-access-management/domain/constants/identity-providers.js';
-import { User } from '../../../../src/identity-access-management/domain/models/User.js';
 import { UserOrganizationForAdmin } from '../../../../src/shared/domain/read-models/UserOrganizationForAdmin.js';
 import * as requestResponseUtils from '../../../../src/shared/infrastructure/utils/request-response-utils.js';
 import { domainBuilder, expect, hFake, sinon } from '../../../test-helper.js';
@@ -217,101 +216,6 @@ describe('Unit | Controller | user-controller', function () {
 
       // then
       expect(response).to.be.equal('ok');
-    });
-  });
-
-  describe('#findPaginatedFilteredUsers', function () {
-    let dependencies;
-
-    beforeEach(function () {
-      sinon.stub(usecases, 'findPaginatedFilteredUsers');
-      const userForAdminSerializer = { serialize: sinon.stub() };
-      dependencies = {
-        userForAdminSerializer,
-      };
-    });
-
-    it('should return a list of JSON API users fetched from the data repository', async function () {
-      // given
-      const request = { query: {} };
-      usecases.findPaginatedFilteredUsers.resolves({ models: {}, pagination: {} });
-      dependencies.userForAdminSerializer.serialize.returns({ data: {}, meta: {} });
-
-      // when
-      await userController.findPaginatedFilteredUsers(request, hFake, dependencies);
-
-      // then
-      expect(usecases.findPaginatedFilteredUsers).to.have.been.calledOnce;
-      expect(dependencies.userForAdminSerializer.serialize).to.have.been.calledOnce;
-    });
-
-    it('should return a JSON API response with pagination information', async function () {
-      // given
-      const request = { query: {} };
-      const expectedResults = [new User({ id: 1 }), new User({ id: 2 }), new User({ id: 3 })];
-      const expectedPagination = { page: 2, pageSize: 25, itemsCount: 100, pagesCount: 4 };
-      usecases.findPaginatedFilteredUsers.resolves({ models: expectedResults, pagination: expectedPagination });
-
-      // when
-      await userController.findPaginatedFilteredUsers(request, hFake, dependencies);
-
-      // then
-      expect(dependencies.userForAdminSerializer.serialize).to.have.been.calledWithExactly(
-        expectedResults,
-        expectedPagination,
-      );
-    });
-
-    it('should allow to filter users by first name', async function () {
-      // given
-      const query = { filter: { firstName: 'Alexia' }, page: {} };
-      const request = { query };
-      usecases.findPaginatedFilteredUsers.resolves({ models: {}, pagination: {} });
-
-      // when
-      await userController.findPaginatedFilteredUsers(request, hFake, dependencies);
-
-      // then
-      expect(usecases.findPaginatedFilteredUsers).to.have.been.calledWithMatch(query);
-    });
-
-    it('should allow to filter users by last name', async function () {
-      // given
-      const query = { filter: { lastName: 'Granjean' }, page: {} };
-      const request = { query };
-      usecases.findPaginatedFilteredUsers.resolves({ models: {}, pagination: {} });
-
-      // when
-      await userController.findPaginatedFilteredUsers(request, hFake, dependencies);
-
-      // then
-      expect(usecases.findPaginatedFilteredUsers).to.have.been.calledWithMatch(query);
-    });
-
-    it('should allow to filter users by email', async function () {
-      // given
-      const query = { filter: { email: 'alexiagranjean' }, page: {} };
-      const request = { query };
-      usecases.findPaginatedFilteredUsers.resolves({ models: {}, pagination: {} });
-
-      // when
-      await userController.findPaginatedFilteredUsers(request, hFake, dependencies);
-
-      // then
-      expect(usecases.findPaginatedFilteredUsers).to.have.been.calledWithMatch(query);
-    });
-
-    it('should allow to paginate on a given page and page size', async function () {
-      // given
-      const query = { filter: { email: 'alexiagranjean' }, page: { number: 2, size: 25 } };
-      const request = { query };
-      usecases.findPaginatedFilteredUsers.resolves({ models: {}, pagination: {} });
-
-      // when
-      await userController.findPaginatedFilteredUsers(request, hFake, dependencies);
-
-      // then
-      expect(usecases.findPaginatedFilteredUsers).to.have.been.calledWithMatch(query);
     });
   });
